pyproject.toml
[project]
name = "apexflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "dbt-bigquery>=1.11.1",
    "google-cloud-bigquery>=3.41.0",
    "google-cloud-storage<3.2",
    "pandas>=3.0.2",
    "pydantic>=2.13.2",
    "requests>=2.33.1",
]
